PI Dialogs: split out base func of LogDialog into more general class
* Most of LogDialog is displaying the contents of a text file, which
  is needed elsewhere (displaying Price quote details...)
